Rework migration safety options (#7)
Create an enum of the valid values of the ``safe`` property.
The current True and False values now map to ``Safe.before_deploy``
and ``Safe.after_deploy``. The default, if unspecified is now
``Safe.after_deploy``. Multiple ``Safe.after_deploy`` migrations
may be dependent on each other without blocking the migration.
Add ``Safe.always`` that can run either before or after the
deployment, so that it runs before the deployment if possible,
but can be delayed until after without blocking the deployment.