Add pre-commit hook (#40)
* Add pre-commit hook

Check the migrations files in the current repository and
ensure that they all have a `safe` property defined.
While we can think of easy ways to break this implementation,
for all the use-cases I've had this covers it quite nicely
so I think it's worth merging even if its rough.

Tim Schilling put an initial sketch of this in
https://github.com/aspiredu/django-safemigrate/issues/39
and I was able to modify it for use in a precommit hook
without difficulty. The modifications were made to tailor
it to the environment of pre-commit, where file names are
passed to the command as arguments. This allows pre-commit
to only need to run the checks incrementally for the affected
files for new commits.

Co-authored-by: Tim Schilling <schillingt@better-simple.com>